Games, topics and stories working

+++ RubioLauraGarciaJuanMartin-Taller4/src/main/java/co/edu/icesi/fi/tics/tssc/delegate/TsscStoryDelegate.java
@@ -6,7 +6,7 @@ public interface TsscStoryDelegate {
 
 	Iterable<TsscStory> findAll();
 	TsscStory saveStory(TsscStory tsscGame);
-	TsscStory findById(int id);
+	TsscStory findById(long id);
 	void delete(TsscStory tsscGame);
 	TsscStory editStory(TsscStory tsscGame);
 }
